refactor folder structure of common crate and deleted ctx.rs in runtime crate
