+++ crates/common/src/wasmtime_runtime/mod.rs
@@ -1,0 +1,5 @@
+mod builder;
+mod linker;
+
+pub use builder::{Builder, WasmtimeBuildable};
+pub use linker::{Linker, WasmtimeLinkable};
